Move tests of input_output_policy into separate file
